requirements.txt
#requirements.txt
torch
torchvision
opencv-python
numpy
pillow
tqdm

